extract keyStepMap into a function that returns the object
